Try to import AES from Cryptodome, use Crypto as a fallback
We need AES_CCM, which is only available in an unreleased version of
Crypto. Moreover, Crypto is technically abandoned.

The drop-in replacement, which does come with AES_CCM, is Cryptodome.
It can be packaged to either use the Crypto namespace (hence being a
drop-in replacement) or use its own Cryptodome namespace. In the
latter case, the package is commonly referred to as cryptodomex.

There is no consensus among Linux distributions on which of these two
approaches is taken (and potentially, both can coexist). To get around
this, we now first try to import AES from Cryptodome and only if that
fails, import it from Crypto. This should work on all distributions.
Hopefully.